Refactor VentaController and models to implement EstadoVenta enum for improved state management and clarity in sales status
